cleanup the repo
